unit testing

+++ bookstore/src/main/java/com/bookstore/app/dto/ProductDTO.java
@@ -2,13 +2,11 @@ package com.bookstore.app.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class ProductDTO {
     private Long productId;
     private String bookTitle;
-    private BigDecimal bookPrice;
+    private double bookPrice;
     private Integer bookQuantity;
     private String bookCategory;
     private String bookDesc;
